[ICRDMA] Fix msan RDMA allocator tests
The allocator is singleton object so we can't place tests with different settins in the one process. The attempt to use non static version of allocator was wrong due to TLS cache inside slot allocator. So right now we just split tests in to 2 process.
